cloud_function/requirements.txt
functions-framework==3.5.0
google-cloud-storage==2.14.0
scikit-learn==1.3.2
numpy==1.26.2
joblib==1.3.2